atc: behavior: fix check build race condition
previously the check factory would create a build and then start it in a
separate database call. if the web node was shut down in between, the
pending build would be left there, but nothing would ever start it, and
checking would just stop until someone comes in and creates a new build.

now the build will be created and started in a single transaction.

a better refactor IMO would be to introduce a `builds.Starter` and use
it to simplify the job build starting and serial groups logic. but i
ain't got time for that!

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>