go-concourse: Return ForbiddenErr from SaveConfig
Add a specific case for 403 errors when saving config (setting
pipeline). This will allow errors such as auth errors or policy check
failures to be handled a bit more explicitly and we can pretty print
their results a bit better. This changes the type of error returned
slightly as ErrForbidden is now of type ForbiddenErr but they are still
both errors so I think this is okay

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>