fly: remove redundant login test
This logic is already tested by the Version Checks suite. There is a
minor difference where the test there implicitly tests the login by
running a different command. And because of that command expects not
only a warning message but a failure. This is a minor difference and is
not worh the slowdown in testing and time outs on CI that this test
comes with.

Signed-off-by: Rui Yang <ryang@pivotal.io>
Co-authored-by: Rui Yang <ryang@pivotal.io>