web/ui: show favorite section for instance groups
we added the favorite section to the instanceGroup view because when
clicking an instance group that's been favorited in the sidebar it takes
you to this view. When the favorite section is not there you then had to
search for your favorite pipeline(s) within the instanc group. With the
favorite section added you no longer need to search (as hard) for the
pipeline.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>