remove zero check order logic
Previously, Concourse would save a resource version with a check order
of 0 using the method SaveUncheckedVersion for get steps that would try
to save metadata on a version that can possibly not exist (because it
got deleted within the time that the version was checked and the get
step finished running). Now, we will save metadata using UpdateMetadata
on a resource which will not update the metadata if the version does not
exist.

Even though we no longer use the SaveUncheckedVersion method, the
logic for it was never deleted. Within this commit, I added a migration
that deletes all rows within the database that has a check order of
zero. This won't affect anything in the users perspective because
versions with check order of 0 are invalid versions that won't even show
up in the web ui. I also removed all queries that have a condition on
check order of 0 because this case should no longer exist anymore.

Signed-off-by: Clara Fu <cfu@pivotal.io>