containerd: use Linux default size for /dev/shm
Guardian defaults to the Linux value which is half of the memory on the
machine. Docker uses 64 MB. For certain use cases, such as running Concourse in
docker-compose in a task, the postgres container requires much more than that.

As mentioned in the link below the memory is not pre-allocated and the
OOM-killer will remove the process if it takes up too much memory.
https://github.com/concourse/concourse/pull/5348#issuecomment-605346946

concourse#6246

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>