atc: add job_id to build image resource caches table
Adding the job id column to the build image resource caches table will
be denormalizing the tables as the information is able to be fetched
from joining to the builds table. Even though it is breaking the current
norm of normalized tables within the database, it greatly improves the
speed of queries so that the queries no longer need to join to the large
builds table.

Signed-off-by: Clara Fu <cfu@pivotal.io>