Implement vault api mocks for api_client testing
This adds support for testing the api_client for vault kv1 and kv2
support. It leverages the ghttp library handlers for this and the
behavior of the AppendHandlers is a bit awkward due to the fact that
each handler is only used once in a specific order. If there is a
better way of registering a route or handler to clean this up id
be happy to refactor a bit.

Signed-off-by: David Dobmeier <david.dobmeier@dcsg.com>