web: behaviour: more lenient status parsing
technically changes behaviour, but not in a meaningful way.  parsing a
filter like "status: pa" will currently result in two pipeline fuzzy
filters - now, it results in a single status filter for the incomplete
status "pa" (which still matches nothing).

the plan is to use Filters in the search bar dropdown logic rather than
raw strings - but the filter parsing needs to be lenient

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>