split base resourcesQuery into two vars
I needed to make a WHERE split around an OR, like this:

SELECT col1 FROM table WHERE (col1 = thing AND col2 = true) OR (col1 = thing2 AND col2 = false)

This is only possible if you're the first Where() called in the query
builder. If you're not the first Where() any constraints you add are
simply AND'd with the previous constrants in the other Where()'s, so I
kept getting this when I tried to add new constraints:

SELECT col1 FROM table WHERE (col1 = thing AND col2 = true) AND (col1 = thing2 AND col2 = false)

By removing the Where() in resourcesQuery I was able to build up the
WHERE how I wanted to.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>