B: render favorited instances in dashboard
with this change, we now display both favourited instance groups AND
favourited pipeline instances in the favourites section (on the
dashboard)

when viewing instance groups (i.e. user applied a `group:...` filter,
either via the search bar or by clicking an instance group), we don't
display the group name in the instance group card, since it appears in
the top-level header and only appears next to other instances in the
same instance group.

when we display fav'd instances in the favorites section, however, these
instances appear adjacent to other favourited pipelines/instance groups,
so we need to display the group name as well (on top of the instance
vars)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>