B: hide "group:" dropdown suggestion in search
since instanced pipelines are still experimental and we need to figure
out the UX for them, it's confusing to suggest the "group:" filter in
the search bar, especially because instanced pipelines are disabled by
default.

with this commit, we hide the "group:" filter from the dropdown so you
don't stumble upon it. note that if you want to apply a "group:" filter,
you still can, and it'll still give suggestions after you've typed
"group:"

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>