atc: add index and mini refactor of check gc
The index will help speed up the deletion of the check gc query in the
check lifecycle. Also refactored the query to use the build_id column on
the resources table to figure out whether or not this build needs to be
gced for a resource. This should help speed up the query a little.

Signed-off-by: Clara Fu <fclara@vmware.com>