B: display instanced pipelines "under" group in sidebar
Rather than displaying favourited instanced pipelines as
"group_name/var1:value1,var2:value2", now we always display the instance
group in the sidebar if there are any favourited instances within it, as
well as all favourited pipeline instances with a "/" as the icon,
omitting the group name. If a favourited instanced pipeline is also
archived, we use the archived icon still (as before).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>