B: fix tooltip for instance vars
fixes a small bug that I introduced in #6453 - hovering over the
instance var name/value would often flicker/not work because it was
triggering mouseover events on the dashboard. using mouseenter instead
and stopping propagation on tooltip elements fixes the problem

note that it only seemed to affect the one  tooltip - others on the
dashboard were fine. either way, stopping propagation on all of them
shouldn't be a problem

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>