B: display all instance vars on the same row
...when in the favourites section, not viewing an instance group.

since favourited instances are displayed alongside "regular" pipelines,
it can look quite odd when an instanced pipeline has many rows since the
regular pipeline has a lot of empty space in the header.

with this change, instanced pipelines are limited to 2 rows when viewed
in the ordinary favorites section (when viewing an instance group, they
can span more rows). instance vars are displayed next to each other,
similar to the set_pipeline step header.

note that this still means there's some empty space on regular
favourited pipelines/instance vars, but when there's just one extra row,
it doesn't look so bad.

this also DRY's the `numHeaderRows` computation in Grid.elm - now, it
uses the same logic as we use to render the rows, so we don't have to
worry about them staying in sync

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>