migration: use go:embed
this was pretty painful; the migration package tests are tightly coupled
to our *actual* migrations all over the place. i converted a few but
then realized just how deep the rabbit hole went, and put in a
hackyRealMigrationsFS function instead.

should anyone want to travel even further down the rabbit hole, there is
a migration package that we wrote and open-sourced but then never found
the time to migrate to - before spending too much time in this package,
we should probably either adopt that package or use some other
off-the-shelf package (if one exists by now):

    https://github.com/concourse/voyager

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>