web: behaviour: Remove worker.Pool lock, add sorting back to "limit-active-tasks"
This removes the lock in worker.Pool. Instead of using a lock to ensure accurate tracking of active tasks, the "limit-active-tasks" strategy optimistically increases the active task count in "Pick" and, if that bumps the number of active tasks over the configured limit, returns an error there (while also decrementing the task count back).

Signed-off-by: David Symons <david@symons.io>